Fix problems caused by iterating over IL
The problem with iterating over the IL is that extra instructions and blocks may be inserted during translation, so it no longer lines up with the x86.

For example, if a block was split because another block branches into the middle of it, an extra LLIL_GOTO will be inserted just before the split, sharing the same address as the instruction before it.

Another problem is with instructions like cmov__, where the IL will expand it into a traditional branch and mov, adding 2 extra blocks, and all the instructions still share the same address.

By iterating over the x86 now, all these problems can be avoided, and the IL can still be used by getting only the first of the instructions with a shared address, which is the only one that matters.