Various fixes (msi, python installer, CMake) - see details.
After I asked an opinion about this, I decided to remove the
GUI launcher.

The python installer (installer.py) should now work correctly
and now supports the --build-msi-package when running under
Windows.

I've added a python script that tests whether the environment
is correctly configured for McSema; in case it is not, the
script attempts to correct the situation by installing what
is missing. This script is located in the start menu.